Move `DefaultHttp{Request,Response}` to `internal.common`
