+++ core/src/main/java/com/linecorp/armeria/client/DecodedHttpResponse.java
@@ -18,9 +18,9 @@ package com.linecorp.armeria.client;
 
 import javax.annotation.Nullable;
 
-import com.linecorp.armeria.common.DefaultHttpResponse;
 import com.linecorp.armeria.common.HttpData;
 import com.linecorp.armeria.common.HttpObject;
+import com.linecorp.armeria.internal.common.DefaultHttpResponse;
 import com.linecorp.armeria.internal.common.InboundTrafficController;
 
 import io.netty.channel.EventLoop;
